+++ tezos/params.go
@@ -40,36 +40,6 @@ var (
 		WithChainId(Ghostnet).
 		WithDeployment(Deployments[Ghostnet].AtProtocol(ProtoV019))
 
-	// NairobinetParams defines the blockchain configuration for Nairobi testnet.
-	// To produce compliant transactions, use these defaults in op.WithParams().
-	NairobinetParams = (&Params{
-		MinimalBlockDelay:            8 * time.Second,
-		CostPerByte:                  250,
-		OriginationSize:              257,
-		HardGasLimitPerOperation:     1040000,
-		HardGasLimitPerBlock:         2600000,
-		HardStorageLimitPerOperation: 60000,
-		MaxOperationDataLength:       32768,
-		MaxOperationsTTL:             240,
-	}).
-		WithChainId(Nairobinet).
-		WithDeployment(Deployments[Nairobinet].AtProtocol(ProtoV017))
-
-	// OxfordnetParams defines the blockchain configuration for Oxford testnet.
-	// To produce compliant transactions, use these defaults in op.WithParams().
-	OxfordnetParams = (&Params{
-		MinimalBlockDelay:            8 * time.Second,
-		CostPerByte:                  250,
-		OriginationSize:              257,
-		HardGasLimitPerOperation:     1040000,
-		HardGasLimitPerBlock:         2600000,
-		HardStorageLimitPerOperation: 60000,
-		MaxOperationDataLength:       32768,
-		MaxOperationsTTL:             240,
-	}).
-		WithChainId(Oxfordnet).
-		WithDeployment(Deployments[Oxfordnet].AtProtocol(ProtoV018))
-
 	// ParisnetParams defines the blockchain configuration for Paris testnet.
 	// To produce compliant transactions, use these defaults in op.WithParams().
 	ParisnetParams = (&Params{
@@ -141,12 +111,10 @@ func (p *Params) WithChainId(id ChainIdHash) *Params {
 			p.Network = "Mainnet"
 		case Ghostnet:
 			p.Network = "Ghostnet"
-		case Nairobinet:
-			p.Network = "Nairobinet"
-		case Oxfordnet:
-			p.Network = "Oxfordnet"
 		case Parisnet:
 			p.Network = "Parisnet"
+		case ParisCnet:
+			p.Network = "ParisCnet"
 		}
 	}
 	return p
@@ -155,7 +123,9 @@ func (p *Params) WithChainId(id ChainIdHash) *Params {
 func (p *Params) WithProtocol(h ProtocolHash) *Params {
 	var ok bool
 	p.Protocol = h
+	versionsMtx.RLock()
 	p.Version, ok = Versions[h]
+	versionsMtx.RUnlock()
 	if !ok {
 		var max int
 		for _, v := range Versions {
@@ -165,6 +135,8 @@ func (p *Params) WithProtocol(h ProtocolHash) *Params {
 			max = v
 		}
 		p.Version = max + 1
+		versionsMtx.Lock()
+		defer versionsMtx.Unlock()
 		Versions[h] = p.Version
 	}
 	switch {
@@ -265,7 +237,8 @@ func (p *Params) CycleFromHeight(height int64) int64 {
 func (p *Params) CycleStartHeight(c int64) int64 {
 	// adjust to target cycle
 	at := p.AtCycle(c)
-	return at.StartHeight - at.StartOffset + (c-at.StartCycle)*at.BlocksPerCycle
+	res := at.StartHeight - at.StartOffset + (c-at.StartCycle)*at.BlocksPerCycle
+	return res
 }
 
 func (p *Params) CycleEndHeight(c int64) int64 {
@@ -295,6 +268,11 @@ func (p *Params) IsCycleEnd(height int64) bool {
 }
 
 func (p *Params) IsSnapshotBlock(height int64) bool {
+	// no more snapshots in Paris
+	if p.Version > 18 && p.IsCycleEnd(height) {
+		return true
+	}
+
 	// adjust to target height
 	at := p.AtBlock(height)
 	pos := at.CyclePosition(height) + 1
@@ -308,7 +286,11 @@ func (p *Params) SnapshotBlock(cycle int64, index int) int64 {
 	if base < 0 {
 		return 0
 	}
-	return at.CycleStartHeight(base) + int64(index+1)*at.BlocksPerSnapshot - 1
+	offset := int64(index+1) * at.BlocksPerSnapshot
+	if offset > at.BlocksPerCycle {
+		offset = at.BlocksPerCycle
+	}
+	return at.CycleStartHeight(base) + offset - 1
 }
 
 func (p *Params) SnapshotIndex(height int64) int {